refactor: reorganized and reshuffled location of graphql typeDefs and resolvers to their own folders
